Move util functions from bls_hsm.h to common
